--- a/core/src/main/scala/caliban/execution/Field.scala
+++ b/core/src/main/scala/caliban/execution/Field.scala
@@ -17,9 +17,11 @@ case class Field(
   fields: List[Field] = Nil,
   conditionalFields: Map[String, List[Field]] = Map(),
   arguments: Map[String, InputValue] = Map(),
-  locationInfo: LocationInfo = LocationInfo.origin,
+  _locationInfo: () => LocationInfo = () => LocationInfo.origin,
   directives: List[Directive] = List.empty
-)
+) {
+  lazy val locationInfo: LocationInfo = _locationInfo()
+}
 
 object Field {
   def apply(
@@ -56,7 +58,7 @@ object Field {
                 field.fields,
                 field.conditionalFields,
                 arguments,
-                sourceMapper.getLocation(index),
+                () => sourceMapper.getLocation(index),
                 directives ++ schemaDirectives
               )
             ),