--- a/core/src/main/scala/caliban/introspection/Introspector.scala
+++ b/core/src/main/scala/caliban/introspection/Introspector.scala
@@ -45,7 +45,7 @@ object Introspector {
       args => types.find(_.name.contains(args.name)).get
     )
     val introspectionSchema = Schema.gen[__Introspection]
-    RootSchema(Operation(introspectionSchema.asType(), introspectionSchema.resolve(resolver)), None, None)
+    RootSchema(Operation(introspectionSchema.toType_(), introspectionSchema.resolve(resolver)), None, None)
   }
 
   private[caliban] def isIntrospection(document: Document): Boolean =