--- a/interop/cats/src/main/scala/caliban/interop/cats/CatsInterop.scala
+++ b/interop/cats/src/main/scala/caliban/interop/cats/CatsInterop.scala
@@ -46,7 +46,7 @@ object CatsInterop {
   def schema[F[_]: Effect, R, A](implicit ev: Schema[R, A]): Schema[R, F[A]] =
     new Schema[R, F[A]] {
       override def toType(isInput: Boolean, isSubscription: Boolean): __Type =
-        ev.asType(isInput, isSubscription)
+        ev.toType_(isInput, isSubscription)
 
       override def optional: Boolean =
         ev.optional