--- a/interop/monix/src/main/scala/caliban/interop/monix/MonixInterop.scala
+++ b/interop/monix/src/main/scala/caliban/interop/monix/MonixInterop.scala
@@ -45,7 +45,7 @@ object MonixInterop {
 
   def taskSchema[R, A](implicit ev: Schema[R, A], ev2: ConcurrentEffect[MonixTask]): Schema[R, MonixTask[A]] =
     new Schema[R, MonixTask[A]] {
-      override def toType(isInput: Boolean, isSubscription: Boolean): __Type = ev.asType(isInput, isSubscription)
+      override def toType(isInput: Boolean, isSubscription: Boolean): __Type = ev.toType_(isInput, isSubscription)
       override def optional: Boolean                                         = ev.optional
       override def resolve(value: MonixTask[A]): Step[R] =
         QueryStep(ZQuery.fromEffect(value.to[Task].map(ev.resolve)))
@@ -57,7 +57,7 @@ object MonixInterop {
     new Schema[R, Observable[A]] {
       override def optional: Boolean = true
       override def toType(isInput: Boolean, isSubscription: Boolean): __Type = {
-        val t = ev.asType(isInput, isSubscription)
+        val t = ev.toType_(isInput, isSubscription)
         if (isSubscription) t else Types.makeList(if (ev.optional) t else Types.makeNonNull(t))
       }
       override def resolve(value: Observable[A]): Step[R] =